+++ aim/web/ui/src/screens/hub/HubMainScreen/components/ContextBox/ContextBox.js
@@ -7,6 +7,7 @@ import Color from 'color';
 import moment from 'moment';
 import * as _ from 'lodash';
 import ContentLoader from 'react-content-loader';
+import { saveAs } from 'file-saver';
 
 import {
   buildUrl,
@@ -16,6 +17,7 @@ import {
   roundValue,
   getCSSSelectorFromString,
   getObjectValueByPath,
+  JSONToCSV,
 } from '../../../../../utils';
 import UI from '../../../../../ui';
 import ContextTable from '../../../../../components/hub/ContextTable/ContextTable';
@@ -23,6 +25,7 @@ import { HUB_PROJECT_EXPERIMENT } from '../../../../../constants/screens';
 import ColumnGroupPopup from './components/ColumnGroupPopup/ColumnGroupPopup';
 import GroupConfigPopup from './components/GroupConfigPopup/GroupConfigPopup';
 import { HubMainScreenModel } from '../../models/HubMainScreenModel';
+import * as analytics from '../../../../../services/analytics';
 
 function ContextBox(props) {
   let [searchFields, setSearchFields] = useState({
@@ -401,6 +404,108 @@ function ContextBox(props) {
     return Object.keys(columnValues).filter(
       (colKey) => columnValues[colKey] !== 'aim_values_differ',
     );
+  }
+
+  function getRowData({ series, chart, runs, paramKeys }) {
+    const { run, metric, trace } = series;
+    const contextHash = contextToHash(trace?.context);
+
+    const line = getTraceData(run.run_hash, metric?.name, contextHash);
+
+    const step = chart.focused.circle.active
+      ? chart.focused.circle.step
+      : chart.focused.step;
+
+    let { stepData } = getClosestStepData(step, line?.data, line?.axisValues);
+
+    let row = {
+      experiment: run.experiment_name ?? '-',
+      run: run.date ? moment.unix(run.date).format('HH:mm · D MMM, YY') : '-',
+      metric: metric?.name ?? '-',
+      context: (() => {
+        const [key, value] = Object.entries(trace.context)[0];
+        return key && value ? `${key}=${value}` : '-';
+      })(),
+      value:
+        stepData !== null && stepData[0] !== null
+          ? roundValue(stepData[0])
+          : '-',
+      step: stepData !== null && stepData[1] !== null ? stepData[1] : '-',
+      epoch: stepData !== null && stepData[2] !== null ? stepData[2] : '-',
+      time:
+        stepData !== null && stepData[3] !== null
+          ? moment.unix(stepData[3]).format('HH:mm:ss · D MMM, YY')
+          : '-',
+    };
+
+    for (let metricKey in runs?.aggMetrics) {
+      runs?.aggMetrics[metricKey].forEach((metricContext) => {
+        row[`${metricKey}-${JSON.stringify(metricContext)}`] = formatValue(
+          series.getAggregatedMetricValue(metricKey, metricContext),
+          true,
+        );
+      });
+    }
+
+    Object.keys(paramKeys).forEach((paramKey) =>
+      paramKeys[paramKey].forEach((key) => {
+        row[`params.${paramKey}.${key}`] = formatValue(
+          run.params?.[paramKey]?.[key],
+        );
+      }),
+    );
+
+    return row;
+  }
+
+  function exportData() {
+    const filteredHeader = Object.keys(table.columnsOrder).reduce(
+      (acc, orderKey) =>
+        acc.concat(
+          table.columnsOrder[orderKey].filter(
+            (column) => table.excludedFields.indexOf(column) === -1,
+          ),
+        ),
+      [],
+    );
+
+    const traceDataToExport = traceList?.traces.reduce(
+      (accArray, traceModel, traceModelIndex) => {
+        (isExploreParamsModeEnabled()
+          ? _.uniqBy(traceModel.series, 'run.run_hash')
+          : traceModel.series
+        ).forEach((series) => {
+          const row = getRowData({
+            series,
+            chart,
+            runs,
+            paramKeys: paramKeys.current,
+          });
+          const filteredRow = filteredHeader.reduce((acc, column) => {
+            acc[column] = row[column];
+            return acc;
+          }, {});
+          accArray.push(filteredRow);
+        });
+
+        if (traceList?.traces.length - 1 !== traceModelIndex) {
+          let emptyRow = {};
+          filteredHeader.forEach((column) => {
+            emptyRow[column] = '--';
+          });
+          accArray.push(emptyRow);
+        }
+        return accArray;
+      },
+      [],
+    );
+
+    const blob = new Blob([JSONToCSV(traceDataToExport)], {
+      type: 'text/csv;charset=utf-8;',
+    });
+    saveAs(blob, `explore-${moment().format('HH:mm:ss · D MMM, YY')}.csv`);
+
+    analytics.trackEvent('[Explore] Export to CSV');
   }
 
   function _renderContentLoader() {
@@ -1343,6 +1448,7 @@ function ContextBox(props) {
             setColumnsOrder={setColumnsOrder}
             columnsWidths={table.columnsWidths}
             setColumnsWidths={setColumnsWidths}
+            exportData={exportData}
             getParamsWithSameValue={getParamsWithSameValue}
             alwaysVisibleColumns={[
               'experiment',